Refactor out Record logic from actual exporter
